refactor: animalCosmetics
